requirements.txt
numpy>=1.19.0
Pillow>=8.0.0
scikit-image>=0.17.2
tqdm>=4.50.0
pycocotools>=2.0.0
scikit-learn>=0.24.0 
ultralytics>=8.0.0
opencv-python>=4.5.0